[rust] Move everything to folders
